root 관련 폴더 구조 변경
- fsd 원칙에 따르면 app 레이어는 응용 프로그램 초기화 계층입니다.
- 기존 App 컴포넌트에서 감싸던 QueryClientProvider와 main.jsx에서 root 요소를 렌더할 때 사용하던 RouterProvider를 관리하는 providers 폴더를 생성했습니다.
- 다음으로 RouterProvider로 전달하는 router를 관리하는 routes 폴더를 생성했습니다.
- 위의 과정을 거쳐 App 컴포넌트는 단순히 화면의 Outlet을 표시하는 기능만을 수행하게 되었습니다.
- NotFound 페이지를 생성했습니다.
