+++ index.html
@@ -9,6 +9,6 @@
   <body>
     <div id="root"></div>
     <div id="modal-root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/app/index.jsx"></script>
   </body>
 </html>
