+++ src/app/index.css
@@ -135,6 +135,14 @@ body {
   outline: none;
 }
 
+html {
+  height: 100%;
+}
+
+#root {
+  height: 100%;
+}
+
 p {
   margin: 0;
   padding: 0;
