+++ src/app/routes/index.js
@@ -1,0 +1,1 @@
+export { router } from "./router";
