--- src/pages/Main.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Main() {
-  return <div>Main</div>;
-}
