--- src/pages/NotFound.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function NotFound() {
-  return <div>NotFound</div>;
-}
