+++ src/pages/not-found/index.js
@@ -1,0 +1,1 @@
+export { default as NotFound } from "./ui/NotFound";
