+++ src/shared/lib/hooks/index.js
@@ -1,3 +1,4 @@
 export { useModal } from "./use-modal";
 export { useModalWithTemp } from "./use-modal";
 export { useNavigateToMain } from "./use-navigate-to-main";
+export { useOutletDisplay } from "./use-outlet-display";
