+++ src/widgets/sign-in-form/api/index.js
@@ -1,3 +1,5 @@
+import { client } from "../../../shared/api";
+
 export async function signIn(data) {
   return await client
     .post("/signin", data)
